Create notify_new_groups.cpp
Not much went into making this program, and eventually I hope to add this as an extension to the add_groups.cpp code in the future. But for now, this works perfectly fine.

Also, the names of .txt files have changed so that they're relatively organized within the Adoptapedia folder on my desktop. See README.cpp for more information.