readWebpage modified for new cases
Case 1) in the event that one of the entries is a dA account and not a
group
Case 2) the watcher value on the page source is written as "###
Watchers" and not as "###Watchers"

Next case to tackle: in the event the webpage does not exist. Future
task: optimize the Case 2 code.